requirements.txt
Flask==2.3.2
opencv-python
mediapipe
numpy
Pillow